create server controls ViewModel
