+++ app/app/src/main/java/me/bickositieff/raspio/ui/serverControls/ServerControlsViewModel.kt
@@ -1,0 +1,16 @@
+package me.bickositieff.raspio.ui.serverControls
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class ServerControlsViewModel : ViewModel() {
+
+    val frequency = MutableLiveData("88")
+
+    val transmissionOn = MutableLiveData(false)
+
+    fun changeTransmissionState(on: Boolean) {
+        transmissionOn.value = on
+    }
+
+}
