--- app/app/src/main/java/me/bickositieff/raspio/ui/servercontrols/ServerControlsViewModel.kt
@@ -1,13 +1,0 @@
-package me.bickositieff.raspio.ui.servercontrols
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class ServerControlsViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
-    }
-    val text: LiveData<String> = _text
-}
