+++ app/app/src/main/res/layout/fragment_server_controls.xml
@@ -1,35 +1,43 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".ui.servercontrols.ServerControlsFragment">
-    <Switch
-            android:text="@string/transmission_state"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content" android:id="@+id/switch1"
-            app:layout_constraintTop_toTopOf="parent"
-            android:layout_marginTop="16dp" app:layout_constraintStart_toStartOf="parent"
-            android:layout_marginStart="16dp" android:textSize="24sp" android:layout_marginEnd="16dp"
-            app:layout_constraintEnd_toEndOf="parent"/>
-    <TextView
-            android:text="@string/frequency"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content" android:id="@+id/textView"
-            app:layout_constraintEnd_toStartOf="@+id/editTextNumber" android:layout_marginEnd="16dp"
-            app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="16dp"
-            android:layout_marginTop="16dp" app:layout_constraintTop_toBottomOf="@+id/switch1" android:textSize="24sp"/>
-    <EditText
-            android:layout_width="179dp"
-            android:layout_height="37dp"
-            android:inputType="number"
-            android:ems="10"
-            android:id="@+id/editTextNumber"
-            app:layout_constraintTop_toBottomOf="@+id/switch1"
-            android:layout_marginTop="16dp"
-            android:layout_marginEnd="16dp"
-            app:layout_constraintEnd_toEndOf="parent" android:textSize="24sp" android:importantForAutofill="no"
-            tools:ignore="LabelFor"/>
-</androidx.constraintlayout.widget.ConstraintLayout>
+<layout xmlns:android="http://schemas.android.com/apk/res/android" xmlns:app="http://schemas.android.com/apk/res-auto"
+        xmlns:tools="http://schemas.android.com/tools">
+    <data>
+        <variable
+                name="control"
+                type="me.bickositieff.raspio.ui.serverControls.ServerControlsViewModel"/>
+    </data>
+    <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            tools:context=".ui.serverControls.ServerControlsFragment">
+        <Switch
+                android:text="@string/transmission_state"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content" android:id="@+id/transmission_state"
+                app:layout_constraintTop_toTopOf="parent"
+                android:layout_marginTop="16dp" app:layout_constraintStart_toStartOf="parent"
+                android:layout_marginStart="16dp" android:textSize="24sp" android:layout_marginEnd="16dp"
+                app:layout_constraintEnd_toEndOf="parent"/>
+        <TextView
+                android:text="@string/frequency"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content" android:id="@+id/frequency_label"
+                app:layout_constraintEnd_toStartOf="@+id/frequency" android:layout_marginEnd="16dp"
+                app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="16dp"
+                android:layout_marginTop="16dp" app:layout_constraintTop_toBottomOf="@+id/transmission_state"
+                android:textSize="24sp"/>
+        <EditText
+                android:enabled="@{!control.transmissionOn}"
+                android:text="@={control.frequency}"
+                android:layout_width="179dp"
+                android:layout_height="37dp"
+                android:inputType="number"
+                android:ems="10"
+                android:id="@+id/frequency"
+                app:layout_constraintTop_toBottomOf="@+id/transmission_state"
+                android:layout_marginTop="16dp"
+                android:layout_marginEnd="16dp"
+                app:layout_constraintEnd_toEndOf="parent" android:textSize="24sp" android:importantForAutofill="no"
+                tools:ignore="LabelFor"/>
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
